Refactor JS for Drag and Drop
