+++ templates/partials/_drop_file_template.html.twig
@@ -1,0 +1,21 @@
+<template id="file-template">
+    <li class="upload-info-item">
+        <div class="icon-file">
+            <i class="fas fa-file-alt"></i>
+        </div>
+        <div class="upload-item-content">
+            <div class="file-name">...</div>
+            <div class="d-flex">
+                <div class="file-size form-text">...</div>
+                <div class="file-error flex-grow-1 form-text text-end"></div>
+            </div>
+            <div class="progress-bar">
+                <div class="progress"></div>
+            </div>
+        </div>
+        <div class="icon-state loading"></div>
+        <div class="icon-remove">
+            <i class="fa-solid fa-xmark"></i>
+        </div>
+    </li>
+</template>
